moved Database to controller and sharedViewModel outside of model
