+++ EventLotterySystemApplication/app/javadocs/com/example/eventlotterysystemapplication/Database.html
@@ -85,7 +85,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class Database" class="title">Class Database</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">com.example.eventlotterysystemapplication.Model.Database</div>
+<div class="inheritance">com.example.eventlotterysystemapplication.Controller.Database</div>
 </div>
 <section class="class-description" id="class-description">
 <hr>
