+++ EventLotterySystemApplication/app/javadocs/com/example/eventlotterysystemapplication/SharedUserViewModel.html
@@ -86,7 +86,7 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance">androidx.lifecycle.ViewModel
-<div class="inheritance">com.example.eventlotterysystemapplication.Model.SharedUserViewModel</div>
+<div class="inheritance">com.example.eventlotterysystemapplication.SharedUserViewModel</div>
 </div>
 </div>
 <section class="class-description" id="class-description">
