+++ EventLotterySystemApplication/app/src/androidTest/java/com/example/eventlotterysystemapplication/DatabaseIntegrationTests.java
@@ -13,7 +13,7 @@ import android.util.Log;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.filters.LargeTest;
 
-import com.example.eventlotterysystemapplication.Model.Database;
+import com.example.eventlotterysystemapplication.Controller.Database;
 import com.example.eventlotterysystemapplication.Model.Event;
 import com.example.eventlotterysystemapplication.Model.User;
 import com.google.firebase.firestore.CollectionReference;
