+++ EventLotterySystemApplication/app/src/main/java/com/example/eventlotterysystemapplication/Model/User.java
@@ -4,6 +4,8 @@ import java.util.Objects;
 
 import android.util.Log;
 
+import com.example.eventlotterysystemapplication.Controller.Database;
+
 /**
  * An instance of this class represents a single user
  */
