+++ EventLotterySystemApplication/app/src/main/java/com/example/eventlotterysystemapplication/SharedUserViewModel.java
@@ -1,8 +1,10 @@
-package com.example.eventlotterysystemapplication.Model;
+package com.example.eventlotterysystemapplication;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
+
+import com.example.eventlotterysystemapplication.Model.User;
 
 // For first time user info fragment to share user info to lottery selection guidelines so committing to database is delayed
 public class SharedUserViewModel extends ViewModel {
