+++ EventLotterySystemApplication/app/src/main/java/com/example/eventlotterysystemapplication/View/CreateOrEditEventFragment.java
@@ -13,7 +13,7 @@ import androidx.annotation.RequiresApi;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.fragment.NavHostFragment;
 
-import com.example.eventlotterysystemapplication.Model.Database;
+import com.example.eventlotterysystemapplication.Controller.Database;
 import com.example.eventlotterysystemapplication.Model.Event;
 import com.example.eventlotterysystemapplication.R;
 import com.example.eventlotterysystemapplication.Model.User;
