+++ EventLotterySystemApplication/app/src/main/java/com/example/eventlotterysystemapplication/View/LotteryGuidelinesFragment.java
@@ -13,9 +13,9 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.navigation.fragment.NavHostFragment;
 
 import com.example.eventlotterysystemapplication.Controller.ContentActivity;
-import com.example.eventlotterysystemapplication.Model.Database;
+import com.example.eventlotterysystemapplication.Controller.Database;
 import com.example.eventlotterysystemapplication.R;
-import com.example.eventlotterysystemapplication.Model.SharedUserViewModel;
+import com.example.eventlotterysystemapplication.SharedUserViewModel;
 import com.example.eventlotterysystemapplication.Model.User;
 import com.example.eventlotterysystemapplication.databinding.FragmentLotteryGuidelinesBinding;
