+++ EventLotterySystemApplication/app/src/test/java/com/example/eventlotterysystemapplication/DatabaseUnitTest.java
@@ -9,7 +9,7 @@ import static org.mockito.Mockito.when;
 
 import android.util.Log;
 
-import com.example.eventlotterysystemapplication.Model.Database;
+import com.example.eventlotterysystemapplication.Controller.Database;
 import com.example.eventlotterysystemapplication.Model.User;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
